refactor: Moved example code into its own location
