+++ src/example/example.h
@@ -1,0 +1,4 @@
+#pragma once
+
+void startExampleApp();
+
